Fix unit tests on MSYS2/Cygwin
POSIX `ln` will create a link even if source is invalid (e.g., does not exist). On Windows platforms, the link will not be created at all. To get around this on Windows, the path and a temporary file are created first, the link is then created, and then the file (and path if also created), are removed.