Fix collection of Windows tests
This is a large set of changes that fixes the majority of remaining test failures on Windows. There are stilll a few stragglers but this gets us quite close. It is a bit more invasive than I'd like, however, as it changes all link tests from '-l' to 'is_symlink' but it could be argued that this also clarifies the code. It may also impact performance but it would be surprising if so and performance testing should be added first before trying to optimize it. Furthermore, the code needs to be entirely cleaned up and all tests fixed before reviewing performance even makes any sense.

* Use normalize_path in tests to ensure results are the same. May switch get_link_target to automatically normalize to reduce changes before pull request.
* Added remove_tree that handles removal of invalid symlinks on Windows.
* Added custom make_path that uses '\\?\' prefix to handle creating paths with dots at the end.
* Replaced all '-l' with 'is_symlink'